+++ src/main/kotlin/trile/Main.kt
@@ -2,11 +2,9 @@ package trile
 
 import io.netty.bootstrap.ServerBootstrap
 import io.netty.channel.ChannelOption
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.atomic.AtomicLong
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.channels.Channel
 import trile.util.addShutdownHook
 import trile.util.createChannelInitializer
 import trile.util.createEventLoopGroup
@@ -16,18 +14,14 @@ fun main() {
 
   val bossGroup = createEventLoopGroup()
   val workerGroup = createEventLoopGroup()
-//  val coroutineScope = CoroutineScope(Executors.newFixedThreadPool(32).asCoroutineDispatcher())
-  val coroutineScope = CoroutineScope(Dispatchers.IO)
-  val capacity = 16384
+//  val dispatcher = Executors.newFixedThreadPool(256).asCoroutineDispatcher()
+  val dispatcher = Dispatchers.IO
+  val coroutineScope = CoroutineScope(dispatcher)
   val counter = AtomicLong()
 
   addShutdownHook(bossGroup, workerGroup)
 
-  val requestChannel = Channel<RequestHolder>(capacity)
-  val requestMap = ConcurrentHashMap<String, RequestHolder>(capacity)
-
-  val handler = CoroutineRequestReceiver(coroutineScope, requestChannel, requestMap, counter)
-  CoroutineRequestProcessor(coroutineScope, requestChannel, requestMap, counter).process()
+  val handler = CoroutineRequestHandler(coroutineScope, counter)
 
   try {
     val server = ServerBootstrap()
